requirements.txt
awscli==1.16.130
boto3==1.9.120
botocore==1.12.120
colorama==0.3.9
configparser==3.7.4
Django==2.2
docutils==0.14
jmespath==0.9.4
mysql-connector==2.2.9
pyasn1==0.4.5
python-dateutil==2.8.0
python-lambda-local==0.1.8
pytz==2018.9
PyYAML==3.13
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
django-cors-headers-2.5.2
djangorestframework==3.9.2
